Fix all the things
