+++ packages/portalnetwork/src/subprotocols/beacon/beacon.ts
@@ -1,14 +1,17 @@
 import { Debugger } from 'debug'
 import { BaseProtocol } from '../protocol.js'
-import { FoundContent } from '../types.js'
-import { ProtocolId } from '../../types.js'
+import { ProtocolId } from '../types.js'
 import { PortalNetwork } from '../../client/client.js'
 import debug from 'debug'
 import { Union } from '@chainsafe/ssz/lib/interface.js'
 import { fromHexString, toHexString } from '@chainsafe/ssz'
 import { shortId } from '../../util/util.js'
 import { createBeaconConfig, defaultChainConfig, BeaconConfig } from '@lodestar/config'
-import { MainnetGenesisValidatorsRoot, BeaconLightClientNetworkContentType } from './types.js'
+import {
+  MainnetGenesisValidatorsRoot,
+  BeaconLightClientNetworkContentType,
+  LightClientUpdatesByRange,
+} from './types.js'
 import {
   ContentMessageType,
   FindContentMessage,
@@ -16,7 +19,7 @@ import {
   PortalWireMessageType,
 } from '../../wire/types.js'
 import { bytesToInt } from '@ethereumjs/util'
-import { RequestCode } from '../../wire/index.js'
+import { RequestCode, FoundContent } from '../../wire/index.js'
 import { ssz } from '@lodestar/types'
 export class BeaconLightClientNetwork extends BaseProtocol {
   protocolId: ProtocolId.BeaconLightClientNetwork
@@ -73,6 +76,7 @@ export class BeaconLightClientNetwork extends BaseProtocol {
     if (res.length === 0) {
       return undefined
     }
+
     try {
       if (bytesToInt(res.subarray(0, 1)) === MessageCodes.CONTENT) {
         this.metrics?.contentMessagesReceived.inc()
