+++ packages/portalnetwork/test/client/provider.spec.ts
@@ -1,6 +1,7 @@
 import { it, assert } from 'vitest'
 import { UltralightProvider } from '../../src/client/provider.js'
-import { ProtocolId, TransportLayer } from '../../src/index.js'
+import { TransportLayer } from '../../src/index.js'
+import { ProtocolId } from '../../src/subprotocols/types.js'
 import { MockProvider } from '../testUtils/mockProvider.js'
 import { Block, BlockHeader } from '@ethereumjs/block'
 import { multiaddr } from '@multiformats/multiaddr'
